extract building date file structure into view container
