+++ src/containers/rx/untagged_container.ts
@@ -1,4 +1,4 @@
-import { App, getAllTags, Menu, moment, TFile } from "obsidian";
+import { App, getAllTags, Menu, TFile } from "obsidian";
 import { ObloggerSettings, ContainerSortMethod, getSortMethodDisplayText, RxGroupType } from "../../settings";
 import { FileState } from "../../constants";
 import { ContainerCallbacks } from "../container_callbacks";
@@ -133,47 +133,6 @@ export class UntaggedContainer extends RxContainer {
                     );
                 }
             });
-    }
-
-    private buildDateFileStructure(
-        unsortedFiles: TFile[],
-        ascending: boolean,
-        useCTime: boolean
-    ) {
-        unsortedFiles.sort((fileA: TFile, fileB: TFile) => {
-            const timestampA = useCTime ? fileA.stat.ctime : fileA.stat.mtime;
-            const timestampB = useCTime ? fileB.stat.ctime : fileB.stat.mtime;
-
-            const monthA = moment(timestampA).format("YYYY-MM");
-            const monthB = moment(timestampB).format("YYYY-MM");
-            if (monthA < monthB) {
-                return ascending ? 1 : -1;
-            } else if (monthA > monthB) {
-                return ascending ? -1 : 1;
-            }
-
-            const bookmarkSorting = this.sortFilesByBookmark(fileA, fileB);
-            if (bookmarkSorting != 0) {
-                return bookmarkSorting;
-            }
-
-            return (ascending ? 1 : -1) * (timestampB - timestampA);
-        }).forEach(file => {
-            const cache = this.app.metadataCache.getFileCache(file);
-            if (cache === null) {
-                console.error("Cache is null after filtering files. This shouldn't happen.");
-                return;
-            }
-            const entryDateString = useCTime ? file.stat.ctime : file.stat.mtime;
-            const entryDate = moment(entryDateString);
-            const entryDateYear = entryDate.format("YYYY");
-            const entryDateMonth = entryDate.format("MM");
-            this.addFileToFolder(
-                file,
-                `${entryDateYear}/${entryDateMonth}`,
-                "/"
-            );
-        });
     }
 
     protected buildFileStructure(excludedFolders: string[]): void {
